+++ pkg/distributor/distributor_test.go
@@ -39,14 +39,12 @@ import (
 	"github.com/weaveworks/common/httpgrpc"
 	"github.com/weaveworks/common/mtime"
 	"github.com/weaveworks/common/user"
-	"go.uber.org/atomic"
 	"golang.org/x/exp/slices"
 	"golang.org/x/time/rate"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/health/grpc_health_v1"
 	"google.golang.org/grpc/status"
 
-	"github.com/grafana/mimir/pkg/distributor/forwarding"
 	"github.com/grafana/mimir/pkg/ingester"
 	"github.com/grafana/mimir/pkg/ingester/client"
 	"github.com/grafana/mimir/pkg/mimirpb"
@@ -413,21 +411,21 @@ func TestDistributor_MetricsCleanup(t *testing.T) {
 		cortex_distributor_received_metadata_total{user="userA"} 5
 		cortex_distributor_received_metadata_total{user="userB"} 10
 
-		# HELP cortex_distributor_received_samples_total The total number of received samples, excluding rejected, forwarded and deduped samples.
+		# HELP cortex_distributor_received_samples_total The total number of received samples, excluding rejected and deduped samples.
 		# TYPE cortex_distributor_received_samples_total counter
 		cortex_distributor_received_samples_total{user="userA"} 5
 		cortex_distributor_received_samples_total{user="userB"} 10
 
-		# HELP cortex_distributor_received_exemplars_total The total number of received exemplars, excluding rejected, forwarded and deduped exemplars.
+		# HELP cortex_distributor_received_exemplars_total The total number of received exemplars, excluding rejected and deduped exemplars.
 		# TYPE cortex_distributor_received_exemplars_total counter
 		cortex_distributor_received_exemplars_total{user="userA"} 5
 		cortex_distributor_received_exemplars_total{user="userB"} 10
 
-		# HELP cortex_distributor_samples_in_total The total number of samples that have come in to the distributor, including rejected, forwarded or deduped samples.
+		# HELP cortex_distributor_samples_in_total The total number of samples that have come in to the distributor, including rejected or deduped samples.
 		# TYPE cortex_distributor_samples_in_total counter
 		cortex_distributor_samples_in_total{user="userA"} 5
 
-		# HELP cortex_distributor_exemplars_in_total The total number of exemplars that have come in to the distributor, including rejected, forwarded or deduped exemplars.
+		# HELP cortex_distributor_exemplars_in_total The total number of exemplars that have come in to the distributor, including rejected or deduped exemplars.
 		# TYPE cortex_distributor_exemplars_in_total counter
 		cortex_distributor_exemplars_in_total{user="userA"} 5
 		`), metrics...))
@@ -451,18 +449,18 @@ func TestDistributor_MetricsCleanup(t *testing.T) {
 		# TYPE cortex_distributor_received_metadata_total counter
 		cortex_distributor_received_metadata_total{user="userB"} 10
 
-		# HELP cortex_distributor_received_samples_total The total number of received samples, excluding rejected, forwarded and deduped samples.
+		# HELP cortex_distributor_received_samples_total The total number of received samples, excluding rejected and deduped samples.
 		# TYPE cortex_distributor_received_samples_total counter
 		cortex_distributor_received_samples_total{user="userB"} 10
 
-		# HELP cortex_distributor_received_exemplars_total The total number of received exemplars, excluding rejected, forwarded and deduped exemplars.
+		# HELP cortex_distributor_received_exemplars_total The total number of received exemplars, excluding rejected and deduped exemplars.
 		# TYPE cortex_distributor_received_exemplars_total counter
 		cortex_distributor_received_exemplars_total{user="userB"} 10
 
-		# HELP cortex_distributor_samples_in_total The total number of samples that have come in to the distributor, including rejected, forwarded or deduped samples.
+		# HELP cortex_distributor_samples_in_total The total number of samples that have come in to the distributor, including rejected or deduped samples.
 		# TYPE cortex_distributor_samples_in_total counter
 
-		# HELP cortex_distributor_exemplars_in_total The total number of exemplars that have come in to the distributor, including rejected, forwarded or deduped exemplars.
+		# HELP cortex_distributor_exemplars_in_total The total number of exemplars that have come in to the distributor, including rejected or deduped exemplars.
 		# TYPE cortex_distributor_exemplars_in_total counter
 		`), metrics...))
 }
@@ -2374,349 +2372,6 @@ func TestDistributor_LabelNamesAndValues_ExpectedAllPossibleLabelNamesAndValuesT
 	require.Equal(t, 10000, len(response.Items[0].Values))
 }
 
-func TestDistributor_IngestionIsControlledByForwarder(t *testing.T) {
-	type testcase struct {
-		name                  string
-		request               *mimirpb.WriteRequest
-		ingestSample          bool
-		expectIngestedMetrics []string
-		expectedMetrics       string
-	}
-
-	metric := "test_metric"
-	testcases := []testcase{
-		{
-			name:                  "do ingest with only samples",
-			request:               makeWriteRequest(123456789000, 5, 0, false, false, metric),
-			ingestSample:          true,
-			expectIngestedMetrics: []string{metric},
-			expectedMetrics: `
-			# HELP cortex_distributor_received_requests_total The total number of received requests, excluding rejected, forwarded and deduped requests.
-			# TYPE cortex_distributor_received_requests_total counter
-			cortex_distributor_received_requests_total{user="user"} 1
-			# HELP cortex_distributor_received_samples_total The total number of received samples, excluding rejected, forwarded and deduped samples.
-			# TYPE cortex_distributor_received_samples_total counter
-			cortex_distributor_received_samples_total{user="user"} 5
-			# HELP cortex_distributor_received_exemplars_total The total number of received exemplars, excluding rejected, forwarded and deduped exemplars.
-			# TYPE cortex_distributor_received_exemplars_total counter
-			cortex_distributor_received_exemplars_total{user="user"} 0
-			# HELP cortex_distributor_received_metadata_total The total number of received metadata, excluding rejected.
-			# TYPE cortex_distributor_received_metadata_total counter
-			cortex_distributor_received_metadata_total{user="user"} 0
-			# HELP cortex_distributor_requests_in_total The total number of requests that have come in to the distributor, including rejected, forwarded or deduped requests.
-			# TYPE cortex_distributor_requests_in_total counter
-			cortex_distributor_requests_in_total{user="user"} 1
-			# HELP cortex_distributor_samples_in_total The total number of samples that have come in to the distributor, including rejected, forwarded or deduped samples.
-			# TYPE cortex_distributor_samples_in_total counter
-			cortex_distributor_samples_in_total{user="user"} 5
-			# HELP cortex_distributor_exemplars_in_total The total number of exemplars that have come in to the distributor, including rejected, forwarded or deduped exemplars.
-			# TYPE cortex_distributor_exemplars_in_total counter
-			cortex_distributor_exemplars_in_total{user="user"} 0
-			# HELP cortex_distributor_metadata_in_total The total number of metadata the have come in to the distributor, including rejected.
-			# TYPE cortex_distributor_metadata_in_total counter
-			cortex_distributor_metadata_in_total{user="user"} 0
-`,
-		}, {
-			name:                  "don't ingest with only samples",
-			request:               makeWriteRequest(123456789000, 5, 0, false, false, metric),
-			ingestSample:          false,
-			expectIngestedMetrics: []string{},
-			expectedMetrics: `
-			# HELP cortex_distributor_received_requests_total The total number of received requests, excluding rejected, forwarded and deduped requests.
-			# TYPE cortex_distributor_received_requests_total counter
-			cortex_distributor_received_requests_total{user="user"} 1
-			# HELP cortex_distributor_received_samples_total The total number of received samples, excluding rejected, forwarded and deduped samples.
-			# TYPE cortex_distributor_received_samples_total counter
-			cortex_distributor_received_samples_total{user="user"} 0
-			# HELP cortex_distributor_received_exemplars_total The total number of received exemplars, excluding rejected, forwarded and deduped exemplars.
-			# TYPE cortex_distributor_received_exemplars_total counter
-			cortex_distributor_received_exemplars_total{user="user"} 0
-			# HELP cortex_distributor_received_metadata_total The total number of received metadata, excluding rejected.
-			# TYPE cortex_distributor_received_metadata_total counter
-			cortex_distributor_received_metadata_total{user="user"} 0
-			# HELP cortex_distributor_requests_in_total The total number of requests that have come in to the distributor, including rejected, forwarded or deduped requests.
-			# TYPE cortex_distributor_requests_in_total counter
-			cortex_distributor_requests_in_total{user="user"} 1
-			# HELP cortex_distributor_samples_in_total The total number of samples that have come in to the distributor, including rejected, forwarded or deduped samples.
-			# TYPE cortex_distributor_samples_in_total counter
-			cortex_distributor_samples_in_total{user="user"} 5
-			# HELP cortex_distributor_exemplars_in_total The total number of exemplars that have come in to the distributor, including rejected, forwarded or deduped exemplars.
-			# TYPE cortex_distributor_exemplars_in_total counter
-			cortex_distributor_exemplars_in_total{user="user"} 0
-			# HELP cortex_distributor_metadata_in_total The total number of metadata the have come in to the distributor, including rejected.
-			# TYPE cortex_distributor_metadata_in_total counter
-			cortex_distributor_metadata_in_total{user="user"} 0
-`,
-		}, {
-			name:                  "do ingest with metadata",
-			request:               makeWriteRequest(123456789000, 5, 5, false, false, metric),
-			ingestSample:          true,
-			expectIngestedMetrics: []string{metric},
-			expectedMetrics: `
-			# HELP cortex_distributor_received_requests_total The total number of received requests, excluding rejected, forwarded and deduped requests.
-			# TYPE cortex_distributor_received_requests_total counter
-			cortex_distributor_received_requests_total{user="user"} 1
-			# HELP cortex_distributor_received_samples_total The total number of received samples, excluding rejected, forwarded and deduped samples.
-			# TYPE cortex_distributor_received_samples_total counter
-			cortex_distributor_received_samples_total{user="user"} 5
-			# HELP cortex_distributor_received_exemplars_total The total number of received exemplars, excluding rejected, forwarded and deduped exemplars.
-			# TYPE cortex_distributor_received_exemplars_total counter
-			cortex_distributor_received_exemplars_total{user="user"} 0
-			# HELP cortex_distributor_received_metadata_total The total number of received metadata, excluding rejected.
-			# TYPE cortex_distributor_received_metadata_total counter
-			cortex_distributor_received_metadata_total{user="user"} 5
-			# HELP cortex_distributor_requests_in_total The total number of requests that have come in to the distributor, including rejected, forwarded or deduped requests.
-			# TYPE cortex_distributor_requests_in_total counter
-			cortex_distributor_requests_in_total{user="user"} 1
-			# HELP cortex_distributor_samples_in_total The total number of samples that have come in to the distributor, including rejected, forwarded or deduped samples.
-			# TYPE cortex_distributor_samples_in_total counter
-			cortex_distributor_samples_in_total{user="user"} 5
-			# HELP cortex_distributor_exemplars_in_total The total number of exemplars that have come in to the distributor, including rejected, forwarded or deduped exemplars.
-			# TYPE cortex_distributor_exemplars_in_total counter
-			cortex_distributor_exemplars_in_total{user="user"} 0
-			# HELP cortex_distributor_metadata_in_total The total number of metadata the have come in to the distributor, including rejected.
-			# TYPE cortex_distributor_metadata_in_total counter
-			cortex_distributor_metadata_in_total{user="user"} 5
-`,
-		}, {
-			name:                  "don't ingest with metadata",
-			request:               makeWriteRequest(123456789000, 5, 5, false, false, metric),
-			ingestSample:          false,
-			expectIngestedMetrics: []string{},
-			expectedMetrics: `
-			# HELP cortex_distributor_received_requests_total The total number of received requests, excluding rejected, forwarded and deduped requests.
-			# TYPE cortex_distributor_received_requests_total counter
-			cortex_distributor_received_requests_total{user="user"} 1
-			# HELP cortex_distributor_received_samples_total The total number of received samples, excluding rejected, forwarded and deduped samples.
-			# TYPE cortex_distributor_received_samples_total counter
-			cortex_distributor_received_samples_total{user="user"} 0
-			# HELP cortex_distributor_received_exemplars_total The total number of received exemplars, excluding rejected, forwarded and deduped exemplars.
-			# TYPE cortex_distributor_received_exemplars_total counter
-			cortex_distributor_received_exemplars_total{user="user"} 0
-			# HELP cortex_distributor_received_metadata_total The total number of received metadata, excluding rejected.
-			# TYPE cortex_distributor_received_metadata_total counter
-			cortex_distributor_received_metadata_total{user="user"} 5
-			# HELP cortex_distributor_requests_in_total The total number of requests that have come in to the distributor, including rejected, forwarded or deduped requests.
-			# TYPE cortex_distributor_requests_in_total counter
-			cortex_distributor_requests_in_total{user="user"} 1
-			# HELP cortex_distributor_samples_in_total The total number of samples that have come in to the distributor, including rejected, forwarded or deduped samples.
-			# TYPE cortex_distributor_samples_in_total counter
-			cortex_distributor_samples_in_total{user="user"} 5
-			# HELP cortex_distributor_exemplars_in_total The total number of exemplars that have come in to the distributor, including rejected, forwarded or deduped exemplars.
-			# TYPE cortex_distributor_exemplars_in_total counter
-			cortex_distributor_exemplars_in_total{user="user"} 0
-			# HELP cortex_distributor_metadata_in_total The total number of metadata the have come in to the distributor, including rejected.
-			# TYPE cortex_distributor_metadata_in_total counter
-			cortex_distributor_metadata_in_total{user="user"} 5
-`,
-		}, {
-			name:                  "do ingest with exemplars",
-			request:               makeWriteRequest(123456789000, 5, 0, true, false, metric),
-			ingestSample:          true,
-			expectIngestedMetrics: []string{metric},
-			expectedMetrics: `
-			# HELP cortex_distributor_received_requests_total The total number of received requests, excluding rejected, forwarded and deduped requests.
-			# TYPE cortex_distributor_received_requests_total counter
-			cortex_distributor_received_requests_total{user="user"} 1
-			# HELP cortex_distributor_received_samples_total The total number of received samples, excluding rejected, forwarded and deduped samples.
-			# TYPE cortex_distributor_received_samples_total counter
-			cortex_distributor_received_samples_total{user="user"} 5
-			# HELP cortex_distributor_received_exemplars_total The total number of received exemplars, excluding rejected, forwarded and deduped exemplars.
-			# TYPE cortex_distributor_received_exemplars_total counter
-			cortex_distributor_received_exemplars_total{user="user"} 5
-			# HELP cortex_distributor_received_metadata_total The total number of received metadata, excluding rejected.
-			# TYPE cortex_distributor_received_metadata_total counter
-			cortex_distributor_received_metadata_total{user="user"} 0
-			# HELP cortex_distributor_requests_in_total The total number of requests that have come in to the distributor, including rejected, forwarded or deduped requests.
-			# TYPE cortex_distributor_requests_in_total counter
-			cortex_distributor_requests_in_total{user="user"} 1
-			# HELP cortex_distributor_samples_in_total The total number of samples that have come in to the distributor, including rejected, forwarded or deduped samples.
-			# TYPE cortex_distributor_samples_in_total counter
-			cortex_distributor_samples_in_total{user="user"} 5
-			# HELP cortex_distributor_exemplars_in_total The total number of exemplars that have come in to the distributor, including rejected, forwarded or deduped exemplars.
-			# TYPE cortex_distributor_exemplars_in_total counter
-			cortex_distributor_exemplars_in_total{user="user"} 5
-			# HELP cortex_distributor_metadata_in_total The total number of metadata the have come in to the distributor, including rejected.
-			# TYPE cortex_distributor_metadata_in_total counter
-			cortex_distributor_metadata_in_total{user="user"} 0
-`,
-		}, {
-			name:                  "don't ingest with exemplars",
-			request:               makeWriteRequest(123456789000, 5, 0, true, false, metric),
-			ingestSample:          false,
-			expectIngestedMetrics: []string{},
-			expectedMetrics: `
-			# HELP cortex_distributor_received_requests_total The total number of received requests, excluding rejected, forwarded and deduped requests.
-			# TYPE cortex_distributor_received_requests_total counter
-			cortex_distributor_received_requests_total{user="user"} 1
-			# HELP cortex_distributor_received_samples_total The total number of received samples, excluding rejected, forwarded and deduped samples.
-			# TYPE cortex_distributor_received_samples_total counter
-			cortex_distributor_received_samples_total{user="user"} 0
-			# HELP cortex_distributor_received_exemplars_total The total number of received exemplars, excluding rejected, forwarded and deduped exemplars.
-			# TYPE cortex_distributor_received_exemplars_total counter
-			cortex_distributor_received_exemplars_total{user="user"} 0
-			# HELP cortex_distributor_received_metadata_total The total number of received metadata, excluding rejected.
-			# TYPE cortex_distributor_received_metadata_total counter
-			cortex_distributor_received_metadata_total{user="user"} 0
-			# HELP cortex_distributor_requests_in_total The total number of requests that have come in to the distributor, including rejected, forwarded or deduped requests.
-			# TYPE cortex_distributor_requests_in_total counter
-			cortex_distributor_requests_in_total{user="user"} 1
-			# HELP cortex_distributor_samples_in_total The total number of samples that have come in to the distributor, including rejected, forwarded or deduped samples.
-			# TYPE cortex_distributor_samples_in_total counter
-			cortex_distributor_samples_in_total{user="user"} 5
-			# HELP cortex_distributor_exemplars_in_total The total number of exemplars that have come in to the distributor, including rejected, forwarded or deduped exemplars.
-			# TYPE cortex_distributor_exemplars_in_total counter
-			cortex_distributor_exemplars_in_total{user="user"} 5
-			# HELP cortex_distributor_metadata_in_total The total number of metadata the have come in to the distributor, including rejected.
-			# TYPE cortex_distributor_metadata_in_total counter
-			cortex_distributor_metadata_in_total{user="user"} 0
-`,
-		}, {
-			name:                  "do ingest with histograms",
-			request:               makeWriteRequest(123456789000, 5, 0, false, true, metric),
-			ingestSample:          true,
-			expectIngestedMetrics: []string{metric},
-			expectedMetrics: `
-			# HELP cortex_distributor_received_requests_total The total number of received requests, excluding rejected, forwarded and deduped requests.
-			# TYPE cortex_distributor_received_requests_total counter
-			cortex_distributor_received_requests_total{user="user"} 1
-			# HELP cortex_distributor_received_samples_total The total number of received samples, excluding rejected, forwarded and deduped samples.
-			# TYPE cortex_distributor_received_samples_total counter
-			cortex_distributor_received_samples_total{user="user"} 10
-			# HELP cortex_distributor_received_exemplars_total The total number of received exemplars, excluding rejected, forwarded and deduped exemplars.
-			# TYPE cortex_distributor_received_exemplars_total counter
-			cortex_distributor_received_exemplars_total{user="user"} 0
-			# HELP cortex_distributor_received_metadata_total The total number of received metadata, excluding rejected.
-			# TYPE cortex_distributor_received_metadata_total counter
-			cortex_distributor_received_metadata_total{user="user"} 0
-			# HELP cortex_distributor_requests_in_total The total number of requests that have come in to the distributor, including rejected, forwarded or deduped requests.
-			# TYPE cortex_distributor_requests_in_total counter
-			cortex_distributor_requests_in_total{user="user"} 1
-			# HELP cortex_distributor_samples_in_total The total number of samples that have come in to the distributor, including rejected, forwarded or deduped samples.
-			# TYPE cortex_distributor_samples_in_total counter
-			cortex_distributor_samples_in_total{user="user"} 10
-			# HELP cortex_distributor_exemplars_in_total The total number of exemplars that have come in to the distributor, including rejected, forwarded or deduped exemplars.
-			# TYPE cortex_distributor_exemplars_in_total counter
-			cortex_distributor_exemplars_in_total{user="user"} 0
-			# HELP cortex_distributor_metadata_in_total The total number of metadata the have come in to the distributor, including rejected.
-			# TYPE cortex_distributor_metadata_in_total counter
-			cortex_distributor_metadata_in_total{user="user"} 0
-`,
-		}, {
-			name:                  "don't ingest with histograms",
-			request:               makeWriteRequest(123456789000, 5, 0, false, true, metric),
-			ingestSample:          false,
-			expectIngestedMetrics: []string{},
-			expectedMetrics: `
-			# HELP cortex_distributor_received_requests_total The total number of received requests, excluding rejected, forwarded and deduped requests.
-			# TYPE cortex_distributor_received_requests_total counter
-			cortex_distributor_received_requests_total{user="user"} 1
-			# HELP cortex_distributor_received_samples_total The total number of received samples, excluding rejected, forwarded and deduped samples.
-			# TYPE cortex_distributor_received_samples_total counter
-			cortex_distributor_received_samples_total{user="user"} 0
-			# HELP cortex_distributor_received_exemplars_total The total number of received exemplars, excluding rejected, forwarded and deduped exemplars.
-			# TYPE cortex_distributor_received_exemplars_total counter
-			cortex_distributor_received_exemplars_total{user="user"} 0
-			# HELP cortex_distributor_received_metadata_total The total number of received metadata, excluding rejected.
-			# TYPE cortex_distributor_received_metadata_total counter
-			cortex_distributor_received_metadata_total{user="user"} 0
-			# HELP cortex_distributor_requests_in_total The total number of requests that have come in to the distributor, including rejected, forwarded or deduped requests.
-			# TYPE cortex_distributor_requests_in_total counter
-			cortex_distributor_requests_in_total{user="user"} 1
-			# HELP cortex_distributor_samples_in_total The total number of samples that have come in to the distributor, including rejected, forwarded or deduped samples.
-			# TYPE cortex_distributor_samples_in_total counter
-			cortex_distributor_samples_in_total{user="user"} 10
-			# HELP cortex_distributor_exemplars_in_total The total number of exemplars that have come in to the distributor, including rejected, forwarded or deduped exemplars.
-			# TYPE cortex_distributor_exemplars_in_total counter
-			cortex_distributor_exemplars_in_total{user="user"} 0
-			# HELP cortex_distributor_metadata_in_total The total number of metadata the have come in to the distributor, including rejected.
-			# TYPE cortex_distributor_metadata_in_total counter
-			cortex_distributor_metadata_in_total{user="user"} 0
-`,
-		},
-	}
-
-	for _, tc := range testcases {
-		t.Run(tc.name, func(t *testing.T) {
-			ctx := user.InjectOrgID(context.Background(), "user")
-			limits := &validation.Limits{
-				ForwardingRules: validation.ForwardingRules{metric: validation.ForwardingRule{}},
-			}
-			flagext.DefaultValues(limits)
-			limits.IngestionRate = 20
-			limits.IngestionBurstSize = 20
-			limits.MaxGlobalExemplarsPerUser = 10
-			limits.ForwardingEndpoint = "non-empty endpoint"
-
-			var forwardReqCnt atomic.Uint32
-			forwardReqCallback := func(_ []mimirpb.PreallocTimeseries) { forwardReqCnt.Inc() }
-			getForwarder := func() forwarding.Forwarder {
-				var mutator func([]mimirpb.PreallocTimeseries) []mimirpb.PreallocTimeseries
-				if tc.ingestSample {
-					mutator = ingestAllTimeseriesMutator
-				}
-				return newMockForwarder(mutator, forwardReqCallback)
-			}
-
-			distributors, ingesters, regs := prepare(t, prepConfig{
-				numIngesters:      1,
-				happyIngesters:    1,
-				replicationFactor: 1,
-				numDistributors:   1,
-				limits:            limits,
-				forwarding:        true,
-				getForwarder:      getForwarder,
-			})
-
-			response, err := distributors[0].Push(ctx, tc.request)
-			assert.NoError(t, err)
-			assert.Equal(t, emptyResponse, response)
-			assert.Equal(t, 1, int(forwardReqCnt.Load()))
-
-			ingestedMetrics := getIngestedMetrics(ctx, t, &ingesters[0])
-			assert.Equal(t, tc.expectIngestedMetrics, ingestedMetrics)
-
-			require.NoError(t, testutil.GatherAndCompare(
-				regs[0],
-				strings.NewReader(tc.expectedMetrics),
-				"cortex_distributor_received_requests_total",
-				"cortex_distributor_received_samples_total",
-				"cortex_distributor_received_exemplars_total",
-				"cortex_distributor_received_metadata_total",
-				"cortex_distributor_requests_in_total",
-				"cortex_distributor_samples_in_total",
-				"cortex_distributor_exemplars_in_total",
-				"cortex_distributor_metadata_in_total",
-			))
-		})
-	}
-}
-
-// getIngestedMetrics takes a mock ingester and returns all the metric names which it has ingested.
-func getIngestedMetrics(ctx context.Context, t *testing.T, ingester *mockIngester) []string {
-	labelsClient, err := ingester.LabelNamesAndValues(ctx, nil)
-	assert.NoError(t, err)
-
-	labels, err := labelsClient.Recv()
-	assert.NoError(t, err)
-
-	resultsUniq := make(map[string]struct{}, len(labels.Items))
-	for _, label := range labels.Items {
-		if label.LabelName == "__name__" {
-			for _, value := range label.Values {
-				resultsUniq[value] = struct{}{}
-			}
-		}
-	}
-
-	results := make([]string, 0, len(resultsUniq))
-	for result := range resultsUniq {
-		results = append(results, result)
-	}
-
-	return results
-}
-
 func prepareWithZoneAwarenessAndZoneDelay(t *testing.T, fixtures []series) (context.Context, []*Distributor) {
 	ctx := user.InjectOrgID(context.Background(), "cardinality-user")
 
@@ -3307,233 +2962,6 @@ func TestRelabelMiddleware(t *testing.T) {
 	}
 }
 
-func TestHaDedupeAndRelabelBeforeForwarding(t *testing.T) {
-	ctx := user.InjectOrgID(context.Background(), "user")
-	const replica1 = "replicaA"
-	const replica2 = "replicaB"
-	const cluster1 = "clusterA"
-	writeReqReplica1 := makeWriteRequestForGenerators(5, labelSetGenForStringPairs(t,
-		"__name__", "foo",
-		"__replica__", replica1, // Will be dropped by ha-dedupe.
-		"bar", "baz", // Will be dropped by drop_label rule.
-		"cluster", cluster1,
-		"sample", "value", // Will be targeted by relabel rule.
-	), nil, nil)
-	writeReqReplica2 := makeWriteRequestForGenerators(5, labelSetGenForStringPairs(t,
-		"__name__", "foo",
-		"__replica__", replica2, // Will be dropped by ha-dedupe.
-		"bar", "baz", // Will be dropped by drop_label rule.
-		"cluster", cluster1,
-		"sample", "value", // Will be targeted by relabel rule.
-	), nil, nil)
-	expectedWriteReq := makeWriteRequestForGenerators(5, labelSetGenForStringPairs(t,
-		"__name__", "foo",
-		"cluster", cluster1,
-		"sample", "value",
-		"target", "prefix_value", // Result of relabel rule.
-	), nil, nil)
-
-	// Capture the submitted write requests which the middlewares pass into the mock push function.
-	var submittedWriteReqs []*mimirpb.WriteRequest
-	mockPush := func(ctx context.Context, pushReq *push.Request) (*mimirpb.WriteResponse, error) {
-		writeReq, err := pushReq.WriteRequest()
-		require.NoError(t, err)
-		submittedWriteReqs = append(submittedWriteReqs, writeReq)
-		return nil, nil
-	}
-
-	// Setup a callback in the mock forwarder to capture the time series which get passed into it by the middleware.
-	var forwardedTs [][]mimirpb.PreallocTimeseries
-	forwardReqCallback := func(ts []mimirpb.PreallocTimeseries) {
-		forwardedTs = append(forwardedTs, ts)
-	}
-	getForwarder := func() forwarding.Forwarder {
-		return newMockForwarder(ingestAllTimeseriesMutator, forwardReqCallback)
-	}
-
-	// Setup limits with HA enabled and forwarding rules for the metric "foo".
-	var limits validation.Limits
-	flagext.DefaultValues(&limits)
-	limits.AcceptHASamples = true
-	limits.MaxLabelValueLength = 15
-	limits.ForwardingEndpoint = "non-empty forwarding endpoint"
-	limits.ForwardingRules = validation.ForwardingRules{
-		"foo": validation.ForwardingRule{},
-	}
-	limits.MetricRelabelConfigs = []*relabel.Config{
-		{
-			SourceLabels: []model.LabelName{"sample"},
-			Action:       relabel.DefaultRelabelConfig.Action,
-			Regex:        relabel.DefaultRelabelConfig.Regex,
-			TargetLabel:  "target",
-			Replacement:  "prefix_$1",
-		},
-	}
-	limits.DropLabels = []string{"bar"}
-
-	// Prepare distributor and wrap the mock push function with its middlewares.
-	ds, _, _ := prepare(t, prepConfig{
-		numDistributors: 1,
-		limits:          &limits,
-		enableTracker:   true,
-		forwarding:      true,
-		getForwarder:    getForwarder,
-	})
-	wrappedMockPush := ds[0].wrapPushWithMiddlewares(mockPush)
-
-	// Submit the two write requests into the wrapped mock push function, it should:
-	// 1) Perform HA-deduplication
-	// 2) Apply relabel rules
-	// 3) Apply drop_label rules
-	// 4) Forward the result via the mock forwarder
-	// 5) Submit the result to the mock push function
-	_, err := wrappedMockPush(ctx, push.NewParsedRequest(writeReqReplica1))
-	assert.NoError(t, err)
-	_, err = wrappedMockPush(ctx, push.NewParsedRequest(writeReqReplica2))
-	resp, ok := httpgrpc.HTTPResponseFromError(err)
-	assert.True(t, ok)
-	assert.Equal(t, int32(202), resp.Code) // 202 due to HA-dedupe.
-
-	// Check that the write requests which have been submitted to the push function look as expected,
-	// there should only be one and it should have all of the expected modifications.
-	assert.Equal(t, []*mimirpb.WriteRequest{expectedWriteReq}, submittedWriteReqs)
-
-	// Check that the time series which have been forwarded via the mock forwarder look as expected,
-	// there should only be one slice of timeseries and it should have all of the expected modifications.
-	assert.Equal(t, [][]mimirpb.PreallocTimeseries{expectedWriteReq.Timeseries}, forwardedTs)
-}
-
-func TestValidationBeforeForwarding(t *testing.T) {
-	ctx := user.InjectOrgID(context.Background(), "user")
-	writeReq := makeWriteRequestForGenerators(5, func(id int) []mimirpb.LabelAdapter {
-		return [][]mimirpb.LabelAdapter{
-			{
-				{
-					Name:  model.MetricNameLabel,
-					Value: "foo",
-				}, {
-					Name:  "some_valid_label1",
-					Value: "some_value",
-				},
-			}, {
-				{
-					Name:  model.MetricNameLabel,
-					Value: "foo",
-				}, {
-					Name:  "some_invalid_label%!?",
-					Value: "some_value",
-				},
-			}, {
-				{
-					Name:  model.MetricNameLabel,
-					Value: "foo",
-				}, {
-					Name:  "this_label_is_too_long_because_it_has_really_a_lot_of_characters",
-					Value: "some_value",
-				},
-			}, {
-				{
-					Name:  model.MetricNameLabel,
-					Value: "foo",
-				}, {
-					Name:  "some_valid_label2",
-					Value: "some_value",
-				},
-			}, {
-				{
-					Name:  model.MetricNameLabel,
-					Value: "foo",
-				}, {
-					Name:  "some_label",
-					Value: "some_value_with_tooooooooo_many_characters",
-				},
-			},
-		}[id]
-	}, nil, nil)
-	expectedWriteReq := makeWriteRequestForGenerators(2, func(id int) []mimirpb.LabelAdapter {
-		return [][]mimirpb.LabelAdapter{
-			{
-				{
-					Name:  model.MetricNameLabel,
-					Value: "foo",
-				}, {
-					Name:  "some_valid_label1",
-					Value: "some_value",
-				},
-			}, {
-				{
-					Name:  model.MetricNameLabel,
-					Value: "foo",
-				}, {
-					Name:  "some_valid_label2",
-					Value: "some_value",
-				},
-			},
-		}[id]
-	}, nil, nil)
-
-	// The values and timestamps are generated by increasing the previous value by 1,
-	// but since validation removes samples 1 and 2 we need to increase the values by 2.
-	expectedWriteReq.Timeseries[1].Samples[0].Value += 2
-	expectedWriteReq.Timeseries[1].Samples[0].TimestampMs += 2
-	// We need to do something similar to samples for histograms, but because the data type has more fields to update it's neater to regenerate it
-	expectedWriteReq.Timeseries[1].Histograms = makeWriteRequestFloatHistograms(103, generateTestFloatHistogram(3))
-
-	// Capture the submitted write requests which the middlewares pass into the mock push function.
-	var submittedWriteReqs []*mimirpb.WriteRequest
-	mockPush := func(ctx context.Context, pushReq *push.Request) (*mimirpb.WriteResponse, error) {
-		writeReq, err := pushReq.WriteRequest()
-		require.NoError(t, err)
-		submittedWriteReqs = append(submittedWriteReqs, writeReq)
-		return nil, nil
-	}
-
-	// Setup a callback in the mock forwarder to capture the time series which get passed into it by the middleware.
-	var forwardedTs [][]mimirpb.PreallocTimeseries
-	forwardReqCallback := func(ts []mimirpb.PreallocTimeseries) {
-		forwardedTs = append(forwardedTs, ts)
-	}
-	getForwarder := func() forwarding.Forwarder {
-		return newMockForwarder(ingestAllTimeseriesMutator, forwardReqCallback)
-	}
-
-	// Setup limits with validation settings which will result in some of the test samples getting filtered.
-	var limits validation.Limits
-	flagext.DefaultValues(&limits)
-	limits.MaxLabelNameLength = 30
-	limits.MaxLabelValueLength = 30
-	limits.ForwardingEndpoint = "non-empty forwarding endpoint"
-	limits.ForwardingRules = validation.ForwardingRules{
-		"foo": validation.ForwardingRule{},
-	}
-
-	// Prepare distributor and wrap the mock push function with its middlewares.
-	ds, _, _ := prepare(t, prepConfig{
-		numDistributors: 1,
-		limits:          &limits,
-		enableTracker:   true,
-		forwarding:      true,
-		getForwarder:    getForwarder,
-	})
-	wrappedMockPush := ds[0].wrapPushWithMiddlewares(mockPush)
-
-	// Submit the write request into the wrapped mock push function,
-	// before samples get forwarded the invalid ones should be removed.
-	_, err := wrappedMockPush(ctx, push.NewParsedRequest(writeReq))
-	assert.ErrorContains(t, err, "received a series with an invalid label")
-	resp, ok := httpgrpc.HTTPResponseFromError(err)
-	require.True(t, ok)
-	assert.Equal(t, int32(400), resp.Code) // 400 because some samples failed validation.
-
-	// Check that the write requests which have been submitted to the push function look as expected,
-	// there should only be one and it should have all of the expected modifications.
-	assert.Equal(t, []*mimirpb.WriteRequest{expectedWriteReq}, submittedWriteReqs)
-
-	// Check that the time series which have been forwarded via the mock forwarder look as expected,
-	// there should only be one slice of timeseries and it should have all of the expected modifications.
-	assert.Equal(t, [][]mimirpb.PreallocTimeseries{expectedWriteReq.Timeseries}, forwardedTs)
-}
-
 func mustNewMatcher(t labels.MatchType, n, v string) *labels.Matcher {
 	m, err := labels.NewMatcher(t, n, v)
 	if err != nil {
@@ -3570,8 +2998,6 @@ type prepConfig struct {
 	ingestersSeriesCountTotal          uint64
 	ingesterZones                      []string
 	labelNamesStreamZonesResponseDelay map[string]time.Duration
-	forwarding                         bool
-	getForwarder                       func() forwarding.Forwarder
 
 	timeOut bool
 }
@@ -3683,12 +3109,6 @@ func prepare(t *testing.T, cfg prepConfig) ([]*Distributor, []mockIngester, []*p
 		distributorCfg.DefaultLimits.MaxIngestionRate = cfg.maxIngestionRate
 		distributorCfg.ShuffleShardingLookbackPeriod = time.Hour
 
-		if cfg.forwarding {
-			distributorCfg.Forwarding.Enabled = true
-			distributorCfg.Forwarding.RequestTimeout = 10 * time.Second
-			distributorCfg.Forwarding.RequestConcurrency = 5
-		}
-
 		cfg.limits.IngestionTenantShardSize = cfg.shuffleShardSize
 
 		if cfg.enableTracker {
@@ -3713,10 +3133,6 @@ func prepare(t *testing.T, cfg prepConfig) ([]*Distributor, []mockIngester, []*p
 		reg := prometheus.NewPedanticRegistry()
 		d, err := New(distributorCfg, clientConfig, overrides, nil, ingestersRing, true, reg, log.NewNopLogger())
 		require.NoError(t, err)
-
-		if cfg.forwarding && cfg.getForwarder != nil {
-			d.forwarder = cfg.getForwarder()
-		}
 
 		require.NoError(t, services.StartAndAwaitRunning(context.Background(), d))
 
@@ -4527,48 +3943,6 @@ outer:
 	return true
 }
 
-type mockForwarder struct {
-	services.Service
-
-	// Optional callback which takes the given timeseries and returns a modified version of it,
-	// the modified version gets returned by Forward().
-	timeseriesMutator func([]mimirpb.PreallocTimeseries) []mimirpb.PreallocTimeseries
-
-	// Optional callback to run in place of the actual forwarding request.
-	forwardReqCallback func([]mimirpb.PreallocTimeseries)
-}
-
-func newMockForwarder(timeseriesMutator func([]mimirpb.PreallocTimeseries) []mimirpb.PreallocTimeseries, forwardReqCallback func([]mimirpb.PreallocTimeseries)) forwarding.Forwarder {
-	return &mockForwarder{
-		timeseriesMutator:  timeseriesMutator,
-		forwardReqCallback: forwardReqCallback,
-	}
-}
-
-func (m *mockForwarder) Forward(ctx context.Context, endpoint string, dontForwardBefore int64, forwardingRules validation.ForwardingRules, ts []mimirpb.PreallocTimeseries, user string) ([]mimirpb.PreallocTimeseries, chan error) {
-	errCh := make(chan error)
-
-	go func() {
-		defer close(errCh)
-
-		if m.forwardReqCallback != nil {
-			m.forwardReqCallback(ts)
-		}
-	}()
-
-	if m.timeseriesMutator != nil {
-		return m.timeseriesMutator(ts), errCh
-	}
-
-	return nil, errCh
-}
-
-func (m *mockForwarder) DeleteMetricsForUser(user string) {}
-
-func ingestAllTimeseriesMutator(ts []mimirpb.PreallocTimeseries) []mimirpb.PreallocTimeseries {
-	return ts
-}
-
 func TestDistributorValidation(t *testing.T) {
 	ctx := user.InjectOrgID(context.Background(), "1")
 	now := model.Now()
@@ -4894,25 +4268,25 @@ func TestDistributor_MetricsWithRequestModifications(t *testing.T) {
 	}
 	getExpectedMetrics := func(cfg expectedMetricsCfg) (string, []string) {
 		return fmt.Sprintf(`
-				# HELP cortex_distributor_requests_in_total The total number of requests that have come in to the distributor, including rejected, forwarded or deduped requests.
+				# HELP cortex_distributor_requests_in_total The total number of requests that have come in to the distributor, including rejected or deduped requests.
 				# TYPE cortex_distributor_requests_in_total counter
 				cortex_distributor_requests_in_total{user="%s"} %d
-				# HELP cortex_distributor_samples_in_total The total number of samples that have come in to the distributor, including rejected, forwarded or deduped samples.
+				# HELP cortex_distributor_samples_in_total The total number of samples that have come in to the distributor, including rejected or deduped samples.
 				# TYPE cortex_distributor_samples_in_total counter
 				cortex_distributor_samples_in_total{user="%s"} %d
-				# HELP cortex_distributor_exemplars_in_total The total number of exemplars that have come in to the distributor, including rejected, forwarded or deduped exemplars.
+				# HELP cortex_distributor_exemplars_in_total The total number of exemplars that have come in to the distributor, including rejected or deduped exemplars.
 				# TYPE cortex_distributor_exemplars_in_total counter
 				cortex_distributor_exemplars_in_total{user="%s"} %d
 				# HELP cortex_distributor_metadata_in_total The total number of metadata the have come in to the distributor, including rejected.
 				# TYPE cortex_distributor_metadata_in_total counter
 				cortex_distributor_metadata_in_total{user="%s"} %d
-				# HELP cortex_distributor_received_requests_total The total number of received requests, excluding rejected, forwarded and deduped requests.
+				# HELP cortex_distributor_received_requests_total The total number of received requests, excluding rejected and deduped requests.
 				# TYPE cortex_distributor_received_requests_total counter
 				cortex_distributor_received_requests_total{user="%s"} %d
-				# HELP cortex_distributor_received_samples_total The total number of received samples, excluding rejected, forwarded and deduped samples.
+				# HELP cortex_distributor_received_samples_total The total number of received samples, excluding rejected and deduped samples.
 				# TYPE cortex_distributor_received_samples_total counter
 				cortex_distributor_received_samples_total{user="%s"} %d
-				# HELP cortex_distributor_received_exemplars_total The total number of received exemplars, excluding rejected, forwarded and deduped exemplars.
+				# HELP cortex_distributor_received_exemplars_total The total number of received exemplars, excluding rejected and deduped exemplars.
 				# TYPE cortex_distributor_received_exemplars_total counter
 				cortex_distributor_received_exemplars_total{user="%s"} %d
 				# HELP cortex_distributor_received_metadata_total The total number of received metadata, excluding rejected.
@@ -5058,42 +4432,6 @@ func TestDistributor_MetricsWithRequestModifications(t *testing.T) {
 			receivedSamples:   40,
 			receivedExemplars: 20,
 			receivedMetadata:  20})
-
-		require.NoError(t, testutil.GatherAndCompare(
-			reg,
-			strings.NewReader(expectedMetrics),
-			metricNames...,
-		))
-	})
-
-	t.Run("Some samples get forwarded", func(t *testing.T) {
-		cfg := getDefaultConfig()
-		cfg.limits.ForwardingEndpoint = "non-empty forwarding endpoint"
-		cfg.limits.ForwardingRules = validation.ForwardingRules{"any_metric": {}} // Needs to be len()>0 to enable forwarding.
-		cfg.forwarding = true
-		cfg.getForwarder = func() forwarding.Forwarder {
-			// This mock forwarder will return half of all samples for ingestion.
-			return newMockForwarder(
-				func(ts []mimirpb.PreallocTimeseries) []mimirpb.PreallocTimeseries {
-					return ts[:len(ts)/2]
-				}, nil,
-			)
-		}
-		dist, reg := getDistributor(cfg)
-		req := makeWriteRequestForGenerators(10, uniqueMetricsGen, exemplarLabelGen, metaDataGen)
-
-		_, err := dist.Push(getCtx(), req)
-		require.NoError(t, err)
-
-		expectedMetrics, metricNames := getExpectedMetrics(expectedMetricsCfg{
-			requestsIn:        1,
-			samplesIn:         20,
-			exemplarsIn:       10,
-			metadataIn:        10,
-			receivedRequests:  1,
-			receivedSamples:   10,
-			receivedExemplars: 5,
-			receivedMetadata:  10})
 
 		require.NoError(t, testutil.GatherAndCompare(
 			reg,